OutputView 에 있던 통계와 결과 산정 로직을 분리
- OutputView 의 통계와 결과 산정 책임을 LottoGame 클래스로 위임하였습니다.
- Main 의 로또번호 생성하는 로직을 LottoController 클래스로 위임하였습니다.
- LottoResult 의 불필요한 속성을 제거하였습니다.
